Menghapus family info dan merubahnya menjadi class
